Format cleanup

+++ ConvertZZ/Moudle/StreamString.cs
@@ -33,7 +33,7 @@ namespace ConvertZZ.Moudle {
             byte[] outBuffer = streamEncoding.GetBytes(outString);
             int len = outBuffer.Length;
             if (len > UInt16.MaxValue) {
-                len = (int)UInt16.MaxValue;
+                len = ushort.MaxValue;
             }
             ioStream.WriteByte((byte)(len / 256));
             ioStream.WriteByte((byte)(len & 255));
@@ -45,7 +45,7 @@ namespace ConvertZZ.Moudle {
             byte[] outBuffer = streamEncoding.GetBytes(outString);
             int len = outBuffer.Length;
             if (len > UInt16.MaxValue) {
-                len = (int)UInt16.MaxValue;
+                len = ushort.MaxValue;
             }
             await ioStream.WriteAsync(new byte[] { (byte)(len / 256) }, 0, 1);
             await ioStream.WriteAsync(new byte[] { (byte)(len & 255) }, 0, 1);
